fix ghost bodies
turns out that GL.DeleteTexture doesnt get rid of the actual texture data, just marks the space as available. On some drivers, this caused a fresh stamp canvas to be created in the same place as the old one (more or less). Clearing the texture before use is the solution apparently lmao